[ARCH-P4] Move query security behind application boundary
